% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.R
\name{rarefy_long}
\alias{rarefy_long}
\title{Rarefaction for long format community data}
\usage{
rarefy_long(x, sites, taxon, counts, sample = NULL, verbose = TRUE)
}
\arguments{
\item{x}{a tibble of community data in long format.}

\item{sites}{column name of the sites.}

\item{taxon}{column name of the taxa.}

\item{counts}{column name of the counts.}

\item{sample}{subsample size for rarefying community.
If \code{NULL} (default) the subsample size
will be the size of the smallest community.}

\item{verbose}{if \code{TRUE} (default), print the subsample size
and the number of sites discarded because below the subsample size.}
}
\value{
a tibble of rarefied community data in long format.
}
\description{
This function can be used to perform rarefaction (subsampling)
of community data.
The actual rarefaction is performed
by \code{\link[vegan]{rrarefy}} from \pkg{vegan}.
}
