% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_zerosum_rows}
\alias{remove_zerosum_rows}
\alias{remove_zerosum_cols}
\title{Remove empty (zero sum) rows and columns}
\usage{
remove_zerosum_rows(x, verbose = TRUE)

remove_zerosum_cols(x, verbose = TRUE)
}
\arguments{
\item{x}{a community matrix or dataframe.}

\item{verbose}{If \code{TRUE} (default), print the names of the rows/columns removed.}
}
\value{
a community matrix or dataframe.
}
\description{
Remove empty (zero sum) rows and columns
}
