% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{save_plots}
\alias{save_plots}
\title{Save multiple plots in one PDF.}
\usage{
save_plots(x, file, width = 8, height = 6, bookmarks = NULL,
  gs.exec = "gs", ...)
}
\arguments{
\item{x}{a list of plots.}

\item{file}{a file path.}

\item{width}{width of the plots.}

\item{height}{height of the plots.}

\item{bookmarks}{a character vector for pdf bookmarks.
If \code{NULL} (default), no bookmarks are added.}

\item{gs.exec}{a path to your Ghostscript executable
(necessary to add bookmarks).}

\item{...}{Other parameters passed on to the pdf function.}
}
\description{
Convenient function for saving multiple plots stored in a list.
}
