% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{spread_cdm}
\alias{spread_cdm}
\title{Long format to community data matrix}
\usage{
spread_cdm(x, row.names, col.names, counts)
}
\arguments{
\item{x}{a tibble.}

\item{row.names}{column name for row names (usually site names).}

\item{col.names}{column name for column names (usually taxonomic names).}

\item{counts}{column name of count data.}
}
\value{
a dataframe in wide format (typical sites x species community data matrix).
}
\description{
Long format to community data matrix
}
