% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tbl_to_df}
\alias{tbl_to_df}
\title{Convert tibbles to dataframes}
\usage{
tbl_to_df(x, row.names)
}
\arguments{
\item{x}{a tibble.}

\item{row.names}{name of the column to use as row names in the returned dataframe.}
}
\value{
a dataframe
}
\description{
Convert tibbles to dataframes
}
