% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{tidy_cdm}
\alias{tidy_cdm}
\title{Community data matrix to tibble}
\usage{
tidy_cdm(x, row.name = "SITE", key.name = "TAXON",
  value.name = "COUNT")
}
\arguments{
\item{x}{a community matrix or dataframe.}

\item{row.name}{name of the column where row names are transfered.}

\item{key.name}{name of the key column.}

\item{value.name}{name of the value column.}
}
\value{
a tibble.
}
\description{
Community data matrix to tibble
}
