% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_plots}
\alias{save_plots}
\title{Save multiple plots in one PDF.}
\usage{
save_plots(.data, ..., files, width = 8, height = 6,
  bookmarks = NULL, gs.exec = "gs")
}
\arguments{
\item{.data}{a tbl.}

\item{...}{one or more list-columns where plots are stored.}

\item{files}{character vector. One file path for each column.}

\item{width}{width of the plots.}

\item{height}{height of the plots.}

\item{bookmarks}{Bookmarks to be added to the PDF. A list of columns generated by vars().
Columns will be interpreted as hierarchical groups and order matters.
Plots will be reodered according to bookmarks in the PDF.
If \code{NULL} (default), no bookmarks are added to the PDF.}

\item{gs.exec}{a path to your Ghostscript executable
(necessary to add bookmarks).}
}
\description{
Convenient function for saving multiple plots stored in a list. The function
can also add bookmarks to the created pdf files.
}
\details{
Bookmarks are added to pdf using Ghostscript, a third party program which
must be installed manually by the user. Tested on Linux only, probably not working
on Windows.
}
