% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spread_cdm}
\alias{spread_cdm}
\title{Long format to community data matrix}
\usage{
spread_cdm(x, row.names, col.names, counts, keep.order = FALSE)
}
\arguments{
\item{x}{a tibble.}

\item{row.names}{column name for row names (usually site names).}

\item{col.names}{column name for column names (usually taxonomic names).}

\item{counts}{column name of count data.}
}
\value{
a dataframe in wide format (typical sites x species community data matrix).
}
\description{
Long format to community data matrix
}
\examples{
x <- data.frame(matrix(rpois(100, 10), nrow = 10))
rownames(x) <- paste("Site", 1:10)
colnames(x) <- paste("Species", LETTERS[1:10])
x
x_tidy <- tidy_cdm(x)
x_tidy
x_df <- spread_cdm(x_tidy, SITE, TAXON, COUNT)
x_df
x_df <- spread_cdm(x_tidy, SITE, TAXON, COUNT, keep.order = TRUE)
all.equal(x, x_df)

}
